oh boy we got controller tracking
controller tracking is now in the project. to actually get it to work is
weird and apparently a unity/hardware problem. so when you have the
whole VR set you need to have the headset semi on and mouse clicking and
holding on the play button, release and put on headset fully at same
time. its really freaking dumb