Player Class features
Player is able to teleport to a Hex that they select.
Player is able to summon both a Big boi and a smol Boi.
these additions have some temp code within it for testing purposes, but
it still includes the proper code for the actual game.
all that remains really for the player class is attacks(fireball and
punch), pick up ball and pass(that will take no time at all), and
interacting with the minions on the field to do things (move, attack).
But with having move and summoning working the others will closely
follow.

ALSO IF YOU HAVE TO CHANGE THE BUILD SETTING PLEASE PUT SOME SORT OF
MESSAGE IN YOU COMMIT PLEASE AND THANKS