issues regarding the player loading in before the game controller and map have beem fixed.
the player now gets the rest of the information after the game
controller and the map have finished loading in.
currently working on fixing an issue with the boiz thats giving me
similar error messages to the player loading in before the map and game
controller. Currently it looks like the minions are not getting the
world map which is an issue. working on a fix