Minion interactions movement and beginning of turn base things
player is now able to interact with the minions to move them. they can
only move a certian amount. started the base for having minions attack.
started the beginning of turn base related things such as the player
being able to end their turn.
other changes include: adding the dome into the game, adding in the
other hex model, some button changes (instead of using left trigger to
open menu it is now the grip button on the left controller, this makes
holding the controller less awkward and uncomfortable)