New Player Interactions
The player now presses the right controller grip button to use the laser
pointer. The player must keep the button pressed down to use, i tried
doing something else like an on off button, but it got weird so i
decided to just make the quicker easier choice for now. if the player
points at a hex that they can move to, a menu pops up on the left
controller of thing to do (move, summon big, summon small) and if you
click the touch pad the action happens. If you point at a ally minion, a
menu of interactions pops up (move, have attack, and pass ball to),
clicking the touch pad lets the action happen. right now there is only
move the minion but it works.

Working on this was a lot of reworking of code that I had already
written. works better now and is more organized. Additional functions to
be added.