fixed little weirdness with the left hand menu and moving a minion
when you were moving a minion to a new hex and you were pointing at
another hex it would give you the hex menu, made it not do that anymore
cause that just makes it confusing.

will add in the future a way to highlight the minion you selected so the
player doesnt forget.